+++ examples/vkcs_kubernetes_addon/base-cluster.tf
@@ -1,0 +1,1 @@
+../vkcs_kubernetes_node_group/main.tf