+++ examples/vkcs_kubernetes_addon/base.tf
@@ -1,0 +1,1 @@
+../vkcs_kubernetes_node_group/base.tf