+++ examples/vkcs_kubernetes_addon/data-addons.tf
@@ -1,0 +1,3 @@
+data "vkcs_kubernetes_addons" "cluster-addons" {
+  cluster_id = vkcs_kubernetes_cluster.k8s-cluster.id
+}