+++ vkcs/internal/services/containerinfraaddons/v1/clusteraddons/urls.go
@@ -1,0 +1,11 @@
+package clusteraddons
+
+import "github.com/gophercloud/gophercloud"
+
+func baseURL() string {
+	return "cluster_addons"
+}
+
+func clusterAddonURL(c *gophercloud.ServiceClient, id string) string {
+	return c.ServiceURL(baseURL(), id)
+}